# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My awesome API",
        "version": "1.0.0"
    },
    "paths": {
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of blog posts",
                "operationId": "App\\Http\\Controllers\\PostsController::index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/posts/{post_id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get blog post by id",
                "description": "Get blog post by id",
                "operationId": "App\\Http\\Controllers\\PostsController::show",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "Post id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Post is not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "text": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}